EcoScanner/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
crypto==1.4.1
cryptography==2.9.2
firebase==3.0.1
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jwcrypto==0.7
lxml==4.5.1
MarkupSafe==1.1.1
Naked==0.1.31
nltk==3.5
oauth2client==4.1.3
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
python-jwt==3.2.6
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
sseclient==0.0.26
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
